requirements.txt
psycopg2-binary
confluent_kafka
simplejson
requests
python-dotenv
pyspark